+++ source_files/EuclidianDistance.cpp
@@ -1,8 +1,8 @@
 #include "Library.hpp"
 
-float EuclidianDistance(const float *vector_1,const float *vector_2,const int vectorSize){
-    float sum = 0.0;
-    float difference = 0.0;
+double EuclidianDistance(const float *vector_1,const float *vector_2,const int vectorSize){
+    double sum = 0.0;
+    double difference = 0.0;
     for(int index = 0; index<vectorSize; index++){
         difference = vector_1[index]-vector_2[index]; // calculate the difference (x0-x1)
         sum+=pow(difference,2); // raise the difference to the power of 2 (x0-x1)^2
